requirements.txt
Flask==3.0.3
dnspython==2.5.0
gunicorn==22.0.0
Werkzeug==3.0.3
